# openapi.yaml
openapi: 3.0.3
info:
  title: Tubes IAE
  description: |
    Dokumentasi API untuk sistem integrasi enterprise retail.

    Sistem terdiri dari 3 aplikasi mandiri (POS, Inventory, CRM)

    **Alur Integrasi End-to-End:**
    1. `POST /transactions` di POS Service → mempublikasikan event `SALE_COMPLETED`
       (Canonical Data Model / CDM) ke RabbitMQ exchange `sales.broadcast`
    2. Integration Service mengonsumsi event, lalu:
       - Mengonversi CDM JSON → SOAP/XML, memanggil `POST /soap/inventory`
       - Memetakan field CDM → CRM, memanggil `POST /customers/update-from-sale`

  version: "1.0.0"
servers:
  - url: http://localhost:8001
    description: POS Service
  - url: http://localhost:8002
    description: Inventory Service
  - url: http://localhost:8003
    description: CRM Service
...
paths:
...
  /crm/customers/{id}:
    get:
      tags: [CRM Service]
      summary: Detail pelanggan
      servers:
        - url: http://localhost:8003
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: CUST-042
      responses:
        "200":
          description: Detail pelanggan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        "404":
          description: Pelanggan tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /crm/customers/update-from-sale:
    post:
      tags: [CRM Service]
      summary: "[Message Translator Target] Update profil pelanggan dari event penjualan"
      description: |
        Endpoint ini **tidak dipanggil langsung oleh client/POS**, melainkan
        oleh **Integration Service** setelah melakukan field mapping dari
        payload CDM ke skema CRM.

        **Field Mapping dari CDM:**

        | CDM (JSON)       | CRM (JSON)     |
        |------------------|----------------|
        | `product_id`     | `item_ref`     |
        | `customer_id`    | `member_code`  |
        | `transaction_id` | `tx_ref`       |
        | `quantity`       | `quantity`     |
        | `unit_price`     | `unit_price`   |

        Poin loyalty dihitung otomatis: 1 poin per Rp 10.000 dari total
        (`quantity * unit_price`).
      servers:
        - url: http://localhost:8003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmUpdateRequest'
            example:
              item_ref: PRD-001
              member_code: CUST-042
              tx_ref: TXN-D20A7935
              quantity: 2
              unit_price: 75000
      responses:
        "200":
          description: Profil pelanggan berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrmUpdateResponse'
              example:
                message: Customer profile updated
                customer:
                  customer_id: CUST-042
                  name: Topek
                  loyalty_points: 30
                  total_purchases: 300000
                points_earned: 15
        "400":
          description: Field wajib tidak lengkap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Product not found
...
    Customer:
      type: object
      properties:
        customer_id:
          type: string
          example: CUST-042
        name:
          type: string
          example: Topek
        loyalty_points:
          type: integer
          example: 30
        total_purchases:
          type: integer
          example: 300000
...
    CrmUpdateRequest:
      type: object
      required: [member_code, tx_ref, quantity, unit_price]
      properties:
        item_ref:
          type: string
          example: PRD-001
        member_code:
          type: string
          example: CUST-042
        tx_ref:
          type: string
          example: TXN-D20A7935
        quantity:
          type: integer
          example: 2
        unit_price:
          type: integer
          example: 75000
    CrmUpdateResponse:
      type: object
      properties:
        message:
          type: string
        customer:
          $ref: '#/components/schemas/Customer'
        points_earned:
          type: integer
          example: 15